job_notification_system/requirements.txt
sqlalchemy==2.0.23
schedule==1.2.0
python-dotenv==1.0.0
jinja2==3.1.2
email-validator==2.1.0
flask==3.0.0